Refactor risk and gui modules for improved calculations and state management
- Updated the RiskArea structure to use f32 types for Dmax, Amax, Amin, and f for better precision in calculations.
- Simplified calculation functions in risk/calculations.zig to return f32 instead of i32, enhancing accuracy.
- Modified draw functions in gui/render.zig to directly use riskProfile properties without unnecessary type conversions.
- Improved text drawing in geo/calculations.zig for better readability.
- Updated gui/state.zig to reflect changes in RiskArea and ensure consistent type usage across the module.

+++ src/risk/ammunition.zig
@@ -23,8 +23,8 @@ pub fn getAmmunitionType(value: i32) Caliber {
 }
 
 pub const Caliber = struct {
-    Dmax: i32,
-    y: i32,
+    Dmax: f32,
+    y: f32,
     c: f32,
     name: []const u8,
